requirements.txt
httpx[http2]
pyyaml
python-dotenv
pandas
plotly
streamlit
